1.引入maven，打包后可直接双击jar包运行(本机有jdk的情况下)。 2.废弃了OrderProperty类，改用Java.util.Properties类来读写config.properties文件。 3.修改了一些判空条件等...
