--- config.properity
@@ -1,10 +1,0 @@
-width=600
-height=480
-username=root
-password=mysql
-database=book
-server_ip=127.0.0.1
-port=3306
-database_state=0
-#网站搜索开关设置，请使用界面上的设置项进行设置
-search_switch=0,1,1,1,1,1
